requirements.txt
FastAPI==0.104.1
uvicorn==0.24.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Paramiko==3.3.1
pycryptodome==3.19.0
requests==2.31.0
geoip2==4.8.0
PyYAML==6.0.1
aiohttp==3.9.1
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
